backend/requirements.txt
# Web Framework (Loosened to allow negotiation)
fastapi
uvicorn[standard]

# Database (Latest version handles dependencies better)
supabase
psycopg2-binary

# AI Integration
google-genai
langchain

# Utilities
python-dotenv
httpx

# Data Validation
pydantic
pydantic-settings

# Testing
pytest
